Remove speed parameter from Ball object
Updating ball rendering and collision detection asynchrony seems to
cause weird bugs (like the ball colliding on nothing when it bounces on
the corner of the screen). This is probably because with async updates
the collision can be detected too late (when the ball already passed the
"safe" X and Y position) so it gets a position where the ball can bounce
forever (or at least until the position gets to a safe place again).

So if we do need to lower requirements to get a acceptable performance,
it's better to just limit the frame rate of rendering (see commit
310d6d0ee55) instead of trying to be clever creating a async update.