Implemented a new way to set the direction of ball after hit the paddle.
We see the paddle as a circumference (actually, as half of a
circumference). If the ball hit the left part of the paddle, it will
move to the left (analogously to the right). The angle of reflection is
proportional to how far from the center the ball hit the paddle.