Add frame skip to Game Loop
On slow devices, it's better to skip some game processing so the game
run slower (and you can see what is happening if your device only
renders few frames per second, instead of seeing the ball jumps like
crazy). The game is pretty light on resources anyway, so no real devices
should trigger this behavior, but this is a small change anyway.